Mild reorg + removal of now-unused code.
